+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 * Fix bug where errors in Main module would show the path of the temporary file instead of the original file
+* Ignore binary files in test directory
 
 # v0.3.0.0